refactor(filter): move FilterPopover back into Filter
